pyproject.toml
[project]
name = "kubernetes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansible>=11.5.0",
    "netaddr>=1.3.0",
    "pyyaml>=6.0.2",
]
